% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Drop Columns with NA}
\usage{
drop_na_columns(df)
}
\arguments{
\item{df}{dataframe as an input}
}
\value{
returns a dataframe without those columns which have na values
}
\description{
\describe{
we've used two methods of handling it and the system handles dataframe of type data.table differently compared to others.
}
\enumerate{
\item for the given dataframe we filter out those columns which are not NA and return the dataframe
\item we choose whether to do splicing or using the Filter function based on the type of dataframe
}
}
