% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{formulate_base_selection_query}
\alias{formulate_base_selection_query}
\title{Base selection query}
\usage{
formulate_base_selection_query(idname, id)
}
\arguments{
\item{idname}{a string denoting the type of id}

\item{id}{a contiguous vector of id values from the array}
}
\value{
a string to use as the input query for scidb to select the given id values
}
\description{
Code to build a query from a list of id input.
\enumerate{
\item formulate_build_literal_query is a function that can be used to create an input query for the
AFL.
\item It takes idname as input.
\item idname is received from get_base_idname under revealgenomics by passing fullarrayname as input
and is the name of the attribute as mentioned in the scidb array.
}
}
