% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{formulate_build_literal_query}
\alias{formulate_build_literal_query}
\title{Code to build a query for sciDB call from the given vec input.}
\usage{
formulate_build_literal_query(vec, value_name = "value_id",
  index_name = "idx")
}
\description{
Code to build a query for sciDB call from the given vec input.
}
