% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{formulate_equi_join_query}
\alias{formulate_equi_join_query}
\title{Equijoin query for scidb}
\usage{
formulate_equi_join_query(left_array_or_query, right_array_or_query,
  left_fields_to_join_by, right_fields_to_join_by,
  keep_dimensions = TRUE)
}
\arguments{
\item{left_array_or_query}{first array to be joined}

\item{right_array_or_query}{second array to be joined}

\item{keep_dimensions}{whether to maintain the dimension ratio of the array}

\item{left_fields_to_join}{the fields from the first array on which to join}

\item{right_fields_to_join}{the fields from the second array on which to join}
}
\value{
returns the query string for performing equijoin in scidb
\describe{
  The formulate_equi_join_query creates the equi join query to be used in scidb based on these parameters
}
}
\description{
equi-join query by taking parameters of both queries and field over which to join them.
}
